🎨 Componentize ListHeading
